Fix: handle focus of menu properly
This commit fixes an issue with the dropdown menu where attempting to click on the new category name input would cause the modal to close.

This was due to the focus of the child not being handled properly. To fix this, we check if the item gaining focus is a child of the menu div, as mentioned in this comment:
https://stackoverflow.com/a/60094794