feat: enable third nav to be assigned from Workspace
Previously, if a user was at their Workspace, they would be able
to create a new page inside a collection, but they would not be able
to assign that page to a third nav in the collection at the same time.
They would have to go through three steps:

1. Create page in collection
2. Navigate to collection
3. Assing said page to third nav section

This commit introduces changes which allow users to reduce these 3
steps to 1, by creating a separate async function which retrieves
the third nav options and loads it into AsyncCreatableSelect.

One strange thing I encountered when attempting to do this was that
when I saved the loadOptions callback as a separate state hook, and updated
that hook with a new function with useEffect whenever `category` changed (to
get the new third nav options for the new `category`), the AsyncCreatableSelect
component would not run the loadOptions function prop for some reason.
I had to define it as a normal function within the ComponentSettingsModal
for it to work. If anyone could figure out why I would be eternally
grateful.