feat: implement new branching logic for file-naming
This commit introduces new branching logic to update the logic to
create file names for the case when a user is trying to create a
collection page inside a third nav section FROM the Workspace. The
conditions are then that:
- `collectionPageData` is null
- `thirdNavTitle` is present

From this, we need to pull the data from the collection structure
endpoint, `${process.env.REACT_APP_BACKEND_URL}/sites/${siteName}/collections/${category}/pages`,
which returns information on third nav sections, instead of the
normal collections endpoint, which only returns information on
collection file names.