fix: bugs with async creatable
Previously, the third nav dropdown would work on the first category
selected, but if a user changed the category selected, the displayed
options would not refresh. This commit fixes that by adding a React
key to the AsyncCreatable component with the value being the current
category. This forces React to rerender it with new props whenever the
category changes.

Next, this commit also uses a hack to save the values of previously
loaded third nav options. Previously, whenever we changed categories,
the dropdown would make the async HTTP request to retrieve third
nav options for that particular category, even if it has been retrieved
before. That's because we clear the cache by setting the key to change
when a category changes.

I tried to set store the retrieved data on a state variable, and to clear
that variable whenever the value of category changes, using a useEffect.
However, since React state is set asynchronously, it doesn't happen in the
right order: the component is re-rendered first, calling the loadOptions
function before the thirdNavData state can be cleared. This means that we
render the previous third nav options instead of the new options that should
be retrieved.

The workaround to this is to store the third nav data as global state,
so that the values are set synchronously.