fix: AsyncCreatableSelect unable to filter options
This commit solves the problem where the AsyncCreatableSelect component
does not filter the options presented depending on the input text. For example,
if I have two third navs, "abc" and "def", and the user types "ab", I want
only third nav options starting with "ab" (meaning, only "abc" and not
"def") to be displayed. This is the default behavior, but somehow
it's not working with AsyncCreatableSelect.

The solution was straightforward - simply import the union function
(default behavior) from the react-select library and use it here. It's
unclear why the default behavior is not working.