feat: display error message if hero button exists but no url is defined
This commit introduces an error messsage which is displayed when
the hero button is defined but the url for the button is not. We are
trying to prevent users from creating a site where clicking on the hero
button does nothing.

This introduces a new block of code to complete this functionality,
instead of updating the existing `validateSections` function. This is
because this error check is dependent on the values of two different
state fields (the hero button and the hero url fields), and we do not
wish to complicated the `validateSections` function with additional
function params.