Fix: Edit Details error
This commit fixes a bug involving the componentSettingsModal being unable to load, which was a bug occasionally experienced when clicking on that option using the dropdown menu.

The reason this was hard to reproduce was due to this bug only occurring if the icon was clicked - the settingsToggle was looking for the id of the item being clicked, and I originally forgot to put an id for the image as well. As a result, the wrong endpoint was being called.

This commit rectifies the problem by re-adding the id to the icon.