fix: allow third navs in a collection to be created from workspace
Previously, we allowed users to create a collection page from the
workspace as well as to assign it to an existing third nav in that
collection. However, this functionality breaks down when the user
specifies a NEW third nav from the workspace, because we do not
distinguish between an existing third nav and a new third nav.

This commit fixes that by distinguishing between the two cases and
generates the correct group identifier.