Fix: change retrieval of allCategories in OverviewCard
This commit fixes how allCategories are retrieved in OverviewCard. Previously, as this retrieval was performed inside OverviewCard, each OverviewCard was performing an identical call to retrieve the the list of categories. In this commit, we move the retrieval into the parent components instead so that this retrieval is only performed once.