fix: prevent save if button text is defined but url is not
Currently users are still allowed to save if the url is empty but
there is button text (happens most frequently if the hero section
is changed from a dropdown to highlights + buttons). This commit
fixes that - we now only display an error message on the url field
instead of the button field.