Fix: update componentSettingsModal to work when new collections are created
This commit fixes a bug where new collections were not being handled correctly. This issue was caused due to 2 reasons - retrieveThirdNavOptions assumed the existence of the collection, and likewise saveFileAndRetrieveUrl also expected the collection to exist, which caused an error to be thrown. This commit fixes both of these util functions.