Fix/edit homepage ui fixes (#230)
* feat: display warning modal if there are any unsaved changes

This commit checks whether there are any unsaved changes on the page
and displays a warning modal if the user tries to navigate away
while there are unsaved changes. This replicates the functionality on
the `EditPage` layout

* feat: display error message if hero button exists but no url is defined

This commit introduces an error messsage which is displayed when
the hero button is defined but the url for the button is not. We are
trying to prevent users from creating a site where clicking on the hero
button does nothing.

This introduces a new block of code to complete this functionality,
instead of updating the existing `validateSections` function. This is
because this error check is dependent on the values of two different
state fields (the hero button and the hero url fields), and we do not
wish to complicated the `validateSections` function with additional
function params.

* feat: make homepage sections chevron more clickable

The designers requested that the click area for the EditHomepage
sections to be larger (to the left). This commit makes that fix by
adding left padding to each of the buttons containing the horizontal
chevron icon so that the click area is larger to the left.

* fix: misc. code cleanup

This commit cleans up code in the following ways:
- replace usage of `class` with `className`
- remove `isRequired` from component PropTypes definitions

* fix: prevent save if button text is defined but url is not

Currently users are still allowed to save if the url is empty but
there is button text (happens most frequently if the hero section
is changed from a dropdown to highlights + buttons). This commit
fixes that - we now only display an error message on the url field
instead of the button field.

* feat: extend button url checking to all sections

This commit extends the null url check for buttons to infobar and
infopic sections as well.

* fix: simplify complicated boolean expression

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>