fix: "Can't perform state update on unmounted component" error
Previously, if you tried to navigate away from a layout / component
while it was still performing an async operation, we would receive
the following error in the console:

```
Warning: Can't perform a React state update on an unmounted component.
This is a no-op, but it indicates a memory leak in your application.
To fix, cancel all subscriptions and asynchronous tasks in a useEffect
cleanup function
```

This is slightly differently-phrased if the component is a class component
but the idea is the same.

The solution to this is to specify a cleanup function. To execute this,
we define an `_isMounted` variable, and only set state updates when
this variable is set to true. In our cleanup function, we set this variable
back to false so that state updates are not performed on unmounted
components.

Reference: https://stackoverflow.com/questions/53949393/cant-perform-a-react-state-update-on-an-unmounted-component