refactor: LoadingButton callback
This commit refactors the LoadingButton so that it runs the
callback in the proper order. Previously, we would attempt to
set the button to a loading state, await on the callback function,
and then set the button back to its non-loading state. The problem
is that the setState functions are not asynchronous so they do not
run on the same event loop as the callback. This commit refactors
the code so that the code is run more reactively using a useEffect.