fix: memory leak in LoadingButton component
Previously, React throws an error complaining about a memory leak
whenever we select an image in the EditPage layout. It turns out
this is because we attempt to set the LoadingButton component's
loading state to false even after the component has unmounted. This
is fixed by applying the same fix from a couple commits ago to the
LoadingButton component.