Feat/add page moving functionality (#219)
* Remove unused component

* Feat: add css for dropdown menu

* Feat: update file saving for when files are moved

* Feat: implement dropdown menu

* Fix: handle focus of menu properly

This commit fixes an issue with the dropdown menu where attempting to click on the new category name input would cause the modal to close.

This was due to the focus of the child not being handled properly. To fix this, we check if the item gaining focus is a child of the menu div, as mentioned in this comment:
https://stackoverflow.com/a/60094794

* Fix: remove outline on focus for dropdown

* Fix: error message display

* Fix: simplify moveFile

* Feat: add icons and improve styling

* Feat: add delete button to dropdown

* Fix: add chevron to move option

* Fix: use loading button for changing category

* Fix: submit correct permalink

* Fix: add keys to list of categoryNames

* Fix: Edit Details error

This commit fixes a bug involving the componentSettingsModal being unable to load, which was a bug occasionally experienced when clicking on that option using the dropdown menu.

The reason this was hard to reproduce was due to this bug only occurring if the icon was clicked - the settingsToggle was looking for the id of the item being clicked, and I originally forgot to put an id for the image as well. As a result, the wrong endpoint was being called.

This commit rectifies the problem by re-adding the id to the icon.

* Fix: close dropdown on opening settings modal

This commit fixes the issue where the dropdown would remain open when the settings modal was opened. To fix this, the button toggling the settings modal now also blurs the focus on the dropdown, to force it to close.

* Fix: change retrieval of allCategories in OverviewCard

This commit fixes how allCategories are retrieved in OverviewCard. Previously, as this retrieval was performed inside OverviewCard, each OverviewCard was performing an identical call to retrieve the the list of categories. In this commit, we move the retrieval into the parent components instead so that this retrieval is only performed once.

* Delete Pages.jsx

* Fix: update componentSettingsModal to work when new collections are created

This commit fixes a bug where new collections were not being handled correctly. This issue was caused due to 2 reasons - retrieveThirdNavOptions assumed the existence of the collection, and likewise saveFileAndRetrieveUrl also expected the collection to exist, which caused an error to be thrown. This commit fixes both of these util functions.

* Fix: make entire row clickable for 'move to' return button

* Fix: change condition for new category creation button to be disabled

Currently, we use `errorMessage` to determine whether the button should be disabled. However, react interprets this as a string and throws an error. To fix this, we use !!errorMessage instead.

* Fix: retrieval of 3rd nav in collection folder

* Fix: update logic for moving files

* Fix: prevent text overflow in menu

* Fix: update regex for category names

* Fix: add validation when creating category

* Fix: form baseApiUrl in saveFileAndRetrieveUrl instead of passing as argument