Fix: reorder deletion and creation of new pages
This commit adjusts the order of file creation/deletion when moving files. Originally, we deleted files before creating a new file, but this caused issues where if problems arose when creating a new file, the old file would still be deleted. This commit changes the deletion to take place after the new file has been created to avoid this loss of data.