Fix: memory leaks in application (#242)
* chore: remove unused Pages layout

* fix: "Can't perform state update on unmounted component" error

Previously, if you tried to navigate away from a layout / component
while it was still performing an async operation, we would receive
the following error in the console:

```
Warning: Can't perform a React state update on an unmounted component.
This is a no-op, but it indicates a memory leak in your application.
To fix, cancel all subscriptions and asynchronous tasks in a useEffect
cleanup function
```

This is slightly differently-phrased if the component is a class component
but the idea is the same.

The solution to this is to specify a cleanup function. To execute this,
we define an `_isMounted` variable, and only set state updates when
this variable is set to true. In our cleanup function, we set this variable
back to false so that state updates are not performed on unmounted
components.

Reference: https://stackoverflow.com/questions/53949393/cant-perform-a-react-state-update-on-an-unmounted-component

* refactor: LoadingButton callback

This commit refactors the LoadingButton so that it runs the
callback in the proper order. Previously, we would attempt to
set the button to a loading state, await on the callback function,
and then set the button back to its non-loading state. The problem
is that the setState functions are not asynchronous so they do not
run on the same event loop as the callback. This commit refactors
the code so that the code is run more reactively using a useEffect.

* fix: memory leak in LoadingButton component

Previously, React throws an error complaining about a memory leak
whenever we select an image in the EditPage layout. It turns out
this is because we attempt to set the LoadingButton component's
loading state to false even after the component has unmounted. This
is fixed by applying the same fix from a couple commits ago to the
LoadingButton component.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>