fix: prevent infinite loading text on Resources layout
Currently on the Resources layout, if the api call to retrieve resource
room name and resource categories fail, then the page continues to display
loading text, which is misleading to users as they'll continue waiting
for a result.

This commit fixes that by setting the load state to be false if there
is an error in the API call.