Fix: add check for save details when toggled
This PR fixes a bug where toggling the dropdown would cause the dropdown options to be toggled as well.

The reason for this bug is an inconsistent behaviour involving onChange on the radio buttons - for some reason, onChange is firing for the Dropdown option even when the actual value is not being changed. To fix this behaviour, we implement an additional check when handling the toggle, to only perform the changes if the sections are being swapped.