feat: add Sentry error boundaries for each route component
This commit adds Sentry error boundaries for each route component
(such as Workspace, Files, etc.). Note that the Sentry error
boundary is simply a wrapper over the built-in React error boundary,
and so does not catch errors for the following events:
- Event handlers
- Asynchronous code (e.g. setTimeout or requestAnimationFrame callbacks)
- Server side rendering
- Errors thrown in the error boundary itself (rather than its children)

Source:
- https://reactjs.org/docs/error-boundaries.html#introducing-error-boundaries
- https://stackoverflow.com/questions/46835391/react-16-error-boundary-component-using-componentdidcatch-shows-uncaught-error

Error boundaries are meant to catch errors during rendering or lifecycle
methods.