Feat/toggle buttons for edit homepage (#257)
* Feat: create handler for toggle

* Feat: update HeroSection to use radio buttons for toggle

* Fix: styling

* Fix: typo causing error with dropdownElems

* Fix: folderCard deleteHandler apiUrl

* Fix: add check for save details when toggled

This PR fixes a bug where toggling the dropdown would cause the dropdown options to be toggled as well.

The reason for this bug is an inconsistent behaviour involving onChange on the radio buttons - for some reason, onChange is firing for the Dropdown option even when the actual value is not being changed. To fix this behaviour, we implement an additional check when handling the toggle, to only perform the changes if the sections are being swapped.

* Fix: standardize hero dropdown format with hero highlights

* Fix: loosen prop type validation requirement

* Fix: initialize dropdownElemErrors to empty array instead of empty string

* Fix: add missing HeroSection proptypes

* Fix: update required proptypes for HeroDropdown

* Fix: radio button value to defaultValue

Previously, we were getting a warning that `A component is changing an uncontrolled input of type radio to be controlled.` The reason for this was that radio buttons should not be using value, but defaultValue instead. This commit makes that change to prevent us from getting that error.

* Fix: pass dropdown and highlights only if they exist to HeroSection