Fix: prevent flickering of dropdown modal
This commit fixes an issue where the dropdown modal would briefly flicker in the wrong position before it could be determined to be out of the viewport. To fix this, we allow the div containing the ref to be initialized first, but prevent the actual menu content from showing up until the position of the menu has been determined.