fix: put section creation buttons outside of droppable area in EditHomepage
This commit resolves issue #259 https://github.com/isomerpages/isomercms-frontend/issues/259.

Previously, when dragging subsections (hero key highlights or hero
dropdown options), when dragging the last element in the section,
the "Create" button would be dragged with it because it existed
within the droppable/draggable area in the component. This commit
fixes that by removing it from that droppable/draggable area.