Feat/add error handling for same file name (#247)
* Fix: reorder deletion and creation of new pages

This commit adjusts the order of file creation/deletion when moving files. Originally, we deleted files before creating a new file, but this caused issues where if problems arose when creating a new file, the old file would still be deleted. This commit changes the deletion to take place after the new file has been created to avoid this loss of data.

* Feat: add new generic AlertModal

* Feat: add error handling in OverviewCard and ComponentSettingsModal

* Fix: close dropdown when alertModal is triggered

* Feat: add warning message when moving files between categories

* Fix: swap button positions on GenericWarningModal

* Install dependencies

* feat: modify GenericWarningModal to allow for customizable button text

* Feat: add ToastContainer to app

* Feat: add Toast component and styling

* Feat: change alert modal to toast notification

* Feat: fix toast styling

* Fix: dropdown modal at the bottom of screen no longer pushes items upwards

* Feat: position dropdown within container if it would exceed to the right

* Nit: rename warning modal hook

* Feat: add propTypes to Toast

* Fix: remove unused AlertModal class

* Fix: prevent flickering of dropdown modal

This commit fixes an issue where the dropdown modal would briefly flicker in the wrong position before it could be determined to be out of the viewport. To fix this, we allow the div containing the ref to be initialized first, but prevent the actual menu content from showing up until the position of the menu has been determined.

* Fix: convert filenames to lowercase when saving

* Feat: add error toast for image and file upload