fix: Create key highlights button
Currently, the "Create highlight" button doesn't work when the
homepage does not have either a dropdown or key highlights hero
section defined. This is possible because we allow users to make
their hero section blank if they do not define any key highlights.

However, this leads to a problem when attempting to create key
highlights in such a situation, because the key highlights array
in the front matter hero section doesn't exist. This leads our code
to attempt to splice something which doesn't exist.

This commit fixes that by creating a new key highlights array in the
front matter hero section if it doesn't already exist.