feat: introduce a new FormFieldDropdown component
There are certain fields in the Settings page, such as show_reach,
or is_government, which should only accept `true` or `false` values.
In such cases, we should limit the user's options instead of allowing
them to enter plain text. This commit introduces a new FormFieldDropdown
component which does just that.