fix: undefined exception when handling errors
When handling errors in some cases, we attempt to check for a http
error response status. However, since the error might not come from
a http error, the `status` attribute might not be present in the error
object, which leads to an undefined attribute exception.

This commit fixes that bug by using optional chaining.