chore: store default error toast message as a constant
This commit modifies all applicable existing usage of the Toast
component to use the default error toast message that is now
defined as a constant in the main utils file. This is so that if
we need to reword our error message, we only need to do it in one
place, instead of in every single file.