Feat/toast notification errors (#281)
* feat: add error toasts when saving and deleting images/files

This commit adds error toast notifications when the api call to
save or delete an image or file fails.

* feat: add error toasts when saving settings

This commit adds error toast notifications when the api call to
save a repo's setting config fails.

* feat: add error toasts for FolderCard

This commit adds error toast notifications when the api call to rename
or to delete a folder (collection or resource category) fails

* feat: add error toasts for OverviewCard

This commit adds error toast notifications when the api call to delete
or to move a page (normal page, collection page, or resource page) fails

* feat: add error toast notifications to ComponentSettingsModal operations

This commit adds error toast notifications when the ComponentSettingsModal
tries but fails to retrieve page setting information. It also modifies the
component so that an error toast is displayed when the api call to save
page settings or to delete the page fails.

* feat: add error toast when we fail to retrieve third nav options

This commit adds an error toast when we fail to retrieve third nav
options in the ComponentSettingsModal

* feat: add error toasts for EditPage and EditHomepage

This commit adds error toasts when the api calls to save EditHomepage
fails, or when the api calls to save or delete the EditPage fails.

* fix: undefined exception when handling errors

When handling errors in some cases, we attempt to check for a http
error response status. However, since the error might not come from
a http error, the `status` attribute might not be present in the error
object, which leads to an undefined attribute exception.

This commit fixes that bug by using optional chaining.

* chore: typo

* chore: store default error toast message as a constant

This commit modifies all applicable existing usage of the Toast
component to use the default error toast message that is now
defined as a constant in the main utils file. This is so that if
we need to reword our error message, we only need to do it in one
place, instead of in every single file.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>