Feat: add access check to ProtectedRoute
This commit adds a check in the ProtectedRoute for whether the user has access to a given site - if the site does not exist or the user does not have the appropriate permissions, they are redirected to a 404 page. This allows us to revert the individual redirections done in most pages, which will be removed in a later commit.