Feat/add 404 and generic error pages (#280)
* Feat: update fallback component with generic error page

* Feat: add 404 page and set direct all unknown urls there

* Feat: add error handling for site pages

This commit fixes the issue of infinitely loading pages when a page does not exist. It also shows the error page if users attempt to edit a page they do not have access to.

* Fix: only show 404 page if backend throws 404 error

* Fix: error in redirect for resource page

* Feat: add access check to ProtectedRoute

This commit adds a check in the ProtectedRoute for whether the user has access to a given site - if the site does not exist or the user does not have the appropriate permissions, they are redirected to a 404 page. This allows us to revert the individual redirections done in most pages, which will be removed in a later commit.

* Fix: revert redirect in individual pages

* Nit: rename variable for redirect

* Fix: remove redirect when page fails to save

The use case covered by these instances on save and delete have now been covered by ProtectedRoute.

* Feat: update NotFoundPage to redirect to workspace if user is on a site

* Feat: add redirect to contact us page