chore: save original settings state
This commit saves the original settings configuration as a state variable
so that we can compare the current settings fields with the original
state to see if any changes were made. This allows us to warn the user
if they attempt to navigate away from the page without saving their
changes.

This commit also removes the `footerSha` attribute from state since
it is no longer required after PR #85 on the backend.