feat: send only settings field which have been changed
This commit modifies the behavior of the Settings layout save handler
so that only fields which have been changed are sent in the request.
Previously, we would send all state fields, regardless of whether the
fields have changed in value. The problem is that this results in
wasteful writes to the GitHub repo on the backend. It also leads to
slow performance since the writes to the GitHub repo are sequential.