Feat/warning if unsaved settings (#288)
* chore: save original settings state

This commit saves the original settings configuration as a state variable
so that we can compare the current settings fields with the original
state to see if any changes were made. This allows us to warn the user
if they attempt to navigate away from the page without saving their
changes.

This commit also removes the `footerSha` attribute from state since
it is no longer required after PR #85 on the backend.

* chore: list out config fields when setting Settings state

* feat: util function for deep comparing objects

This commit introduces a util function i found on stack overflow,
https://stackoverflow.com/a/39813128, which deep compares two objects and
identifies which attributes have changed.

* feat: display warnings if there are unsaved changes on settings page

This commit deep compares the current settings values with the original
loaded settings values and displays a warning if the user attempts to
navigate away from the page when there are unsaved changes.

* feat: send only settings field which have been changed (#286)

This commit modifies the behavior of the Settings layout save handler
so that only fields which have been changed are sent in the request.
Previously, we would send all state fields, regardless of whether the
fields have changed in value. The problem is that this results in
wasteful writes to the GitHub repo on the backend. It also leads to
slow performance since the writes to the GitHub repo are sequential.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>

* chore: code cleanup

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>