feat: add util function for modifying EditPage color theme
This commit adds a util function, createPageStyleSheet, which
creates a custom style sheet based on the primary and secondary colors
of a given repo. This will allow us to customize the color appearance
of the EditPage live preview.

For now, this function only modifies the html elements associated with
the EditPage layout. This will be expanded to the EditContactUs and
EditHomepage layouts in the future.

This commit also introduces a getSiteColors function which retrieves
a repo's settings file and retrieves the primary and secondary colors
for that repo.