feat: introduce siteColors state hook on App level
This commit introduces a `siteColors` state hook object at the App
level. This state keeps track of all the primary and secondary colors
of each repo, and the `siteColors` state hook and its respective
setter function are passed as props to every route.

This commit also makes modifications to the Workspace layout so that
we retrieve a repo's color settings when we reach the Workspace layout,
only if we have not already done so before.

Ideally, we would have used a React context to store the siteColors
object instead of a state hook at the App level. However, since some
of the components (EditPage, EditHomepage, EditContactUs) are class
components, we are not able to use the React Context API unless we
refactor these class components into hooks, which would take significant
engineering effort.