feat: modify EditPage color scheme
This commit modifies the EditPage mounting process. If the site
colors for the repo are already available, we simply generate the
custom style sheet and apply it to the layout, which will give us
the new color scheme.

If the site colors are not yet available, we make an API call to retrieve
them and subsequently apply them to the layout.

We do not implement a toast notification if the call to retrieve
the color settings fail - we simply let it fail silently. This is because
the core function of the page is to display page content, and the color
of the preview is only a secondary function. However, if anyone else
feels strongly otherwise we can discuss and implement a toast notification
instead.