feat: update resources section in the EditHomepage layout
This commit updates the resources section in the EditHomepage layout
by modifying the watermark to include a more informative message which
lets the user know that the resource tiles displayed are just samples.

The styles of the watermark have also been updated, and the opacity of the
resource tiles decreased so that the message is in greater contrast
and more readable.