Feat/use site colors (#296)
* feat: experiment - please ignore this commit

This commit introduces the `primaryColors` and `secondaryColors`
state hook at the application root, `App.jsx`. By keeping track of
colors at the root level, we can supply this colors to the `EditPage`
components directly, so we only need to retrieve settings data for
each repository once.

The state hook is stored as an object, where the key is the repo name
and the value is the hex color string. The reason for this is so that
we can keep track of the color of each repo. If we used the state hook
to store only the current repo's color, if the user redirected to another
repo, then that repo's `EditPage` would be rendered with another repo's
colors, which would just be wrong.

We also modify the `EditPage` component to retrieve color settings
data ONLY if the state hook does not have color data for that repo.
No data is fetched if the state hook already has color data for that
repo. Therefore, for each repo, we only need to fetch color data once.
If the user opens up another page on `EditPage`, we would already have
color data for that repo and so we can send it as a prop.

* feat: add util function for modifying EditPage color theme

This commit adds a util function, createPageStyleSheet, which
creates a custom style sheet based on the primary and secondary colors
of a given repo. This will allow us to customize the color appearance
of the EditPage live preview.

For now, this function only modifies the html elements associated with
the EditPage layout. This will be expanded to the EditContactUs and
EditHomepage layouts in the future.

This commit also introduces a getSiteColors function which retrieves
a repo's settings file and retrieves the primary and secondary colors
for that repo.

* feat: introduce siteColors state hook on App level

This commit introduces a `siteColors` state hook object at the App
level. This state keeps track of all the primary and secondary colors
of each repo, and the `siteColors` state hook and its respective
setter function are passed as props to every route.

This commit also makes modifications to the Workspace layout so that
we retrieve a repo's color settings when we reach the Workspace layout,
only if we have not already done so before.

Ideally, we would have used a React context to store the siteColors
object instead of a state hook at the App level. However, since some
of the components (EditPage, EditHomepage, EditContactUs) are class
components, we are not able to use the React Context API unless we
refactor these class components into hooks, which would take significant
engineering effort.

* feat: modify EditPage color scheme

This commit modifies the EditPage mounting process. If the site
colors for the repo are already available, we simply generate the
custom style sheet and apply it to the layout, which will give us
the new color scheme.

If the site colors are not yet available, we make an API call to retrieve
them and subsequently apply them to the layout.

We do not implement a toast notification if the call to retrieve
the color settings fail - we simply let it fail silently. This is because
the core function of the page is to display page content, and the color
of the preview is only a secondary function. However, if anyone else
feels strongly otherwise we can discuss and implement a toast notification
instead.

* feat: update style sheet creation function for EditHomepage

This commit updates the createPageStyleSheet util function to
account for HTML elements on the EditHomepage layout

* feat: update EditHomepage to implement custom style sheet

This commit implements a custom style sheet so that the live preview's
colors match that of the actual site. This is done in the same way
that we previously did for the EditPage layout.

* fix: minor bug in custom style sheet creation

* chore: prevent memory leaks for EditContactUs

* feat: implement custom styles for EditContactUs

* fix: equality check

* feat: update resources section in the EditHomepage layout

This commit updates the resources section in the EditHomepage layout
by modifying the watermark to include a more informative message which
lets the user know that the resource tiles displayed are just samples.

The styles of the watermark have also been updated, and the opacity of the
resource tiles decreased so that the message is in greater contrast
and more readable.

* fix: align resource placeholder text to center

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>