fix: remove use of enquote string (#311)
Currently, when saving data for resource pages, we enquote the title
in anticipation of special characters in the title. However, this turns
out to be unnecessary, and in fact, detrimental because our yaml library
automatically enquotes strings which contains special characters. This
manual enquoting results in a title string with two double quotes.

This commit fixes that by removing the use of enquote string and letting
the library do its work.

Co-authored-by: Jie Hao Kwa <jiehaokwa@Jies-MacBook-Pro.local>