fix: unable to select same file twice for upload
Steps to reproduce bug:
1. Go to the Files layout
2. Click on upload file
3. Select a file. At this point, the file renaming modal should pop up.
4. Exit the file renaming modal
5. Click on upload file again
6. Select the same file
7. Nothing happens - the file renaming modal does not show up

The reason this bug occurs is that the onload event that our filereader
detects for changes to the file input value. When there is no change
to the value, the `fileReader.onload` event doesn't get triggered.

The fix is to reset the file input value after every use.

For some reason, this bug does not affect image uploads.